requirements.txt
beautifulsoup4==4.11.1
fake_useragent==0.1.14
htmldate==1.3.2
numpy==1.23.4
pandas==1.5.1
python-dotenv==0.21.0
requests==2.25.1
selenium==4.6.0
webdriver_manager==3.8.4
